python/flowforge-otel/pyproject.toml
[project]
name = "flowforge-otel"
version = "0.1.0"
description = "flowforge-otel — OpenTelemetry-backed TracingPort + HistogramMetricsPort adapter"
license = "Apache-2.0"
license-files = ["LICENSE"]
authors = [{ name = "Nyimbi Odero, Datacraft", email = "nyimbi@gmail.com" }]
maintainers = [{ name = "Nyimbi Odero, Datacraft", email = "nyimbi@gmail.com" }]
keywords = ["workflow", "opentelemetry", "tracing", "metrics", "observability"]
classifiers = [
	"Development Status :: 4 - Beta",
	"Intended Audience :: Developers",
	"Intended Audience :: Information Technology",
	"License :: OSI Approved :: Apache Software License",
	"Operating System :: OS Independent",
	"Programming Language :: Python",
	"Programming Language :: Python :: 3",
	"Programming Language :: Python :: 3.11",
	"Programming Language :: Python :: 3.12",
	"Programming Language :: Python :: 3.13",
	"Topic :: Software Development :: Libraries",
	"Topic :: Software Development :: Libraries :: Python Modules",
	"Topic :: System :: Distributed Computing",
	"Topic :: System :: Monitoring",
	"Framework :: AsyncIO",
	"Framework :: FastAPI",
	"Framework :: Pydantic",
	"Framework :: Pydantic :: 2",
	"Typing :: Typed",
]
# pypi-metadata-stamped

readme = "README.md"
requires-python = ">=3.11"
dependencies = [
	"flowforge",
]

[project.optional-dependencies]
otel = [
	"opentelemetry-api>=1.20",
	"opentelemetry-sdk>=1.20",
]
test = [
	"pytest>=7",
	"pytest-asyncio>=0.23",
	"opentelemetry-api>=1.20",
	"opentelemetry-sdk>=1.20",
]

[project.urls]
Homepage = "https://github.com/nyimbi/ums"
Documentation = "https://github.com/nyimbi/ums/tree/main/framework/docs"
Repository = "https://github.com/nyimbi/ums"
Issues = "https://github.com/nyimbi/ums/issues"
Changelog = "https://github.com/nyimbi/ums/blob/main/framework/CHANGELOG.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/flowforge_otel"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
